requirements.txt
efficientnet_pytorch==0.7.1
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.4.0
Pillow==9.5.0
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.65.0
